Csv translating feature
